Add setting up instructions and assets
We are going with Atom rather than VS Code, as VS Code is an IDE
and by nature has more features than required. This might become
an issue if students have laptops that are low on resources.

Atom is a text editor. With a text editor, students are required to
learn things that an IDE would do for you.